+++ internal/model/balance.go
@@ -1,7 +1,7 @@
 package model
 
 type Balance struct {
-	ID        int64 `json:"id"`
-	Current   int64 `json:"current"`
-	Withdrawn int64 `json:"withdrawn"`
+	ID        int64   `json:"id"`
+	Current   float64 `json:"current"`
+	Withdrawn float64 `json:"withdrawn"`
 }
